ZCS-16660 UpdateD SOAP API Documentation for SendMsg, SaveDraft and SecureSendMsg
Indentation
