+++ soap/src/java/com/zimbra/soap/mail/message/SendMsgRequest.java
@@ -105,13 +105,6 @@ public class SendMsgRequest {
     @XmlElement(name=MailConstants.E_MSG /* m */, required=false)
     private MsgToSend msg;
 
-    /**
-     * @zm-api-field-tag delivery-receipt-notification
-     * @zm-api-field-description If set, delivery receipt notification will be sent.
-     */
-    @XmlAttribute(name=MailConstants.A_DELIVERY_RECEIPT_NOTIFICATION /* deliveryReport */, required=false)
-    private ZmBoolean deliveryReport;
-
     public SendMsgRequest() {
     }
 
@@ -131,10 +124,6 @@ public class SendMsgRequest {
         this.fetchSavedMsg = ZmBoolean.fromBool(fetchSavedMsg);
     }
 
-    public void setDeliveryReport(Boolean deliveryReport) {
-        this.deliveryReport = ZmBoolean.fromBool(deliveryReport);
-    }
-
     public void setSendUid(String sendUid) { this.sendUid = sendUid; }
     public void setMsg(MsgToSend msg) { this.msg = msg; }
 
@@ -143,7 +132,6 @@ public class SendMsgRequest {
     public Boolean getNoSaveToSent() { return ZmBoolean.toBool(noSaveToSent); }
     public Boolean getFetchSavedMsg() { return ZmBoolean.toBool(fetchSavedMsg); }
     public String getSendUid() { return sendUid; }
-    public Boolean getDeliveryReport() { return ZmBoolean.toBool(deliveryReport); }
     public MsgToSend getMsg() { return msg; }
 
     public MoreObjects.ToStringHelper addToStringInfo(MoreObjects.ToStringHelper helper) {
@@ -153,7 +141,6 @@ public class SendMsgRequest {
             .add("noSaveToSent", noSaveToSent)
             .add("fetchSavedMsg", fetchSavedMsg)
             .add("sendUid", sendUid)
-            .add("deliveryReport", deliveryReport)
             .add("msg", msg);
     }
